admin için bundle oluşturme fixlendi artık kurslar ile birlikte bundle oluşturması gerek
admin için oluşturduğu bundleları getirme api'ı yazıldı

bundlelar getirilirken Kurs sınıfı da dahil edildi

remaining kurs sayfası için progressler getirildi
kalan kurs sayısı 0 değilse kurs gösterildi

bu apiler için gerekli olan dtolar hem response hem request oluşturuldu

bu apiler için gerekli olan mapperlar oluşturuldu

+++ src/main/java/com/example/EtiyaSportsClub/dtos/BundlesByUserId.java
@@ -14,7 +14,7 @@ import java.sql.Timestamp;
 public class BundlesByUserId {
     private Long bundleId;
     private String bundleName;
-    private int bundlePrice;
+    private double bundlePrice;
     private String bundleDescription;
     private int totalLessonNumber;
     private int remainingCourseNumber;
